Begin process of creating UI
currenlty in the process of laying out the UI
Saddly in this loop of devlopment we will not be able to create a UI that completely matches the design document.
simply because i do not know how to devlop that well in java. i could learn how to improve on this but since we are using agile right now i want to focus on getting a working project.
also keeping in mind time factors (we need to get this out the door next week)
i will include this in the documentation of the assessment